Update search tracking
- Normalise query string to account for legacy site url encoding
- Switch tracking to event tracking rather than pageviews
